[FEAT] Clone del servidor

+++ git-server/tests/server_test.rs
@@ -8,19 +8,26 @@ use std::{
 #[ignore = "Needs server"]
 fn test_push() {
     let path = "./tests/data/test1";
-    let git_bin = "../../../../target/debug/git";
 
-    create_base_scene(path.clone());
-
-    assert!(
-        Command::new("git")
-            .arg("clone")
-            .arg("git://127.1.0.0:9418/repo")
-            .arg("repo")
-            .current_dir(path)
-            .status()
-            .is_ok(),
-        "No se pudo agregar el archivo testfile"
+    let rebuild_server_repo = false;
+    create_base_scene(path.clone(), rebuild_server_repo);
+    if rebuild_server_repo {
+        panic!();
+    }
+    let result = Command::new("git")
+        .arg("clone")
+        .arg("git://127.1.0.0:9418/repo")
+        .arg("repo")
+        .current_dir(path)
+        .output()
+        .unwrap();
+    println!(
+        "===\nstderr: {}\n===",
+        String::from_utf8(result.stderr).unwrap()
+    );
+    println!(
+        "===\nstdout: {}\n===",
+        String::from_utf8(result.stdout).unwrap()
     );
 
     let mut readme = File::open(path.to_owned() + "/repo/README.md").unwrap();
@@ -28,15 +35,47 @@ fn test_push() {
     readme.read_to_string(&mut contents).unwrap();
     assert_eq!(contents, "Commit inicial\n");
 
-    panic!("STOP");
-
-    let mut file = File::create(path.to_owned() + "/repo/testfile2").unwrap();
-    file.write_all(b"contenido2\n").unwrap();
+    let mut file = File::create(path.to_owned() + "/repocan/testfile").unwrap();
+    file.write_all(b"contenido\n").unwrap();
 
     assert!(
         Command::new("../".to_string() + "git")
             .arg("add")
-            .arg("testfile2")
+            .arg("testfile")
+            .current_dir(path.to_owned() + "/repocan")
+            .status()
+            .is_ok(),
+        "No se pudo agregar el archivo testfile"
+    );
+
+    assert!(
+        Command::new("../".to_string() + "git")
+            .arg("commit")
+            .arg("-m")
+            .arg("hi2")
+            .current_dir(path.to_owned() + "/repocan")
+            .status()
+            .is_ok(),
+        "No se pudo hacer commit"
+    );
+
+    assert!(
+        Command::new("../".to_owned() + "git")
+            .arg("push")
+            .current_dir(&format!("{}/repocan", path))
+            .status()
+            .is_ok(),
+        "No se pudo agregar el archivo testfile"
+    );
+
+    panic!("STOP");
+    let mut file = File::create(path.to_owned() + "/repo/testfile").unwrap();
+    file.write_all(b"contenido\n").unwrap();
+
+    assert!(
+        Command::new("../".to_string() + "git")
+            .arg("add")
+            .arg("testfile")
             .current_dir(path.to_owned() + "/repo")
             .status()
             .is_ok(),
@@ -87,10 +126,12 @@ fn test_push() {
     _ = fs::remove_dir_all(path);
 }
 
-fn create_base_scene(path: &str) {
-    // _ = fs::remove_dir_all(format!("{}/server-files/repo", path));
+fn create_base_scene(path: &str, rebuild_server_repo: bool) {
+    if rebuild_server_repo {
+        _ = fs::remove_dir_all(format!("{}/server-files/repo", path));
+        _ = fs::remove_dir_all(format!("{}/repocan", path));
+    }
     _ = fs::remove_dir_all(format!("{}/repo", path));
-    _ = fs::remove_dir_all(format!("{}/repocan", path));
 
     let Ok(_) = fs::create_dir_all(path.clone()) else {
         panic!("No se pudo crear el directorio")
@@ -99,9 +140,7 @@ fn create_base_scene(path: &str) {
     let Ok(_) = fs::create_dir_all(format!("{}/server-files/repo", path)) else {
         panic!("No se pudo crear el directorio")
     };
-    let Ok(_) = fs::create_dir_all(format!("{}/repocan", path)) else {
-        panic!("No se pudo crear el directorio")
-    };
+
     let Ok(_) = fs::create_dir_all(format!("{}/other_user", path)) else {
         panic!("No se pudo crear el directorio")
     };
@@ -138,46 +177,48 @@ fn create_base_scene(path: &str) {
         "No se pudo crear el archivo git-daemon-export-ok"
     );
 
-    // assert!(
-    //     Command::new("git")
-    //         .arg("clone")
-    //         .arg("git://127.1.0.0:9418/repo")
-    //         .arg("repocan")
-    //         .current_dir(path.to_owned())
-    //         .status()
-    //         .is_ok(),
-    //     "No se pudo clonar con canónico"
-    // );
-    // let mut file = File::create(path.to_owned() + "/repocan/README.md").unwrap();
-    // file.write_all(b"Commit inicial\n").unwrap();
+    if rebuild_server_repo {
+        assert!(
+            Command::new("git")
+                .arg("clone")
+                .arg("git://127.1.0.0:9418/repo")
+                .arg("repocan")
+                .current_dir(path.to_owned())
+                .status()
+                .is_ok(),
+            "No se pudo clonar con canónico"
+        );
+        let mut file = File::create(path.to_owned() + "/repocan/README.md").unwrap();
+        file.write_all(b"Commit inicial\n").unwrap();
 
-    // assert!(
-    //     Command::new("git")
-    //         .arg("add")
-    //         .arg("README.md")
-    //         .current_dir(path.to_owned() + "/repocan")
-    //         .status()
-    //         .is_ok(),
-    //     "No se pudo agregar el archivo testfile"
-    // );
+        assert!(
+            Command::new("git")
+                .arg("add")
+                .arg("README.md")
+                .current_dir(path.to_owned() + "/repocan")
+                .status()
+                .is_ok(),
+            "No se pudo agregar el archivo testfile"
+        );
 
-    // assert!(
-    //     Command::new("git")
-    //         .arg("commit")
-    //         .arg("-m")
-    //         .arg("InitialCommit")
-    //         .current_dir(path.to_owned() + "/repocan")
-    //         .status()
-    //         .is_ok(),
-    //     "No se pudo hacer commit"
-    // );
+        assert!(
+            Command::new("git")
+                .arg("commit")
+                .arg("-m")
+                .arg("InitialCommit")
+                .current_dir(path.to_owned() + "/repocan")
+                .status()
+                .is_ok(),
+            "No se pudo hacer commit"
+        );
 
-    // assert!(
-    //     Command::new("git")
-    //         .arg("push")
-    //         .current_dir(path.to_owned() + "/repocan")
-    //         .status()
-    //         .is_ok(),
-    //     "No se pudo hacer push"
-    // );
+        assert!(
+            Command::new("git")
+                .arg("push")
+                .current_dir(path.to_owned() + "/repocan")
+                .status()
+                .is_ok(),
+            "No se pudo hacer push"
+        );
+    }
 }
